Move pkg/subscription to pkg/customizations
